move readers in new `io.py` module and ipdate imports
